refactor: (tobe) 이름 Value Object 통합
