--- src/main/java/kitchenpos/menus/domain/tobe/domain/menugroup/Name.java
@@ -1,43 +1,0 @@
-package kitchenpos.menus.domain.tobe.domain.menugroup;
-
-import java.util.Objects;
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class Name {
-
-    @Column(name = "name", nullable = false)
-    private String value;
-
-    protected Name() {
-    }
-
-    public Name(final String value) {
-        validate(value);
-        this.value = value;
-    }
-
-    private void validate(final String value) {
-        if (Objects.isNull(value) || value.isEmpty()) {
-            throw new IllegalArgumentException("메뉴 그룹의 이름이 올바르지 않으면 등록할 수 없습니다.");
-        }
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        final Name name = (Name) o;
-        return Objects.equals(value, name.value);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(value);
-    }
-}
