+++ src/main/java/kitchenpos/products/domain/tobe/domain/Product.java
@@ -5,6 +5,7 @@ import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
+import kitchenpos.common.domain.DisplayedName;
 import kitchenpos.common.domain.Price;
 import kitchenpos.common.domain.ProductId;
