+++ src/test/java/kitchenpos/common/domain/DisplayedNameTest.java
@@ -1,10 +1,9 @@
-package kitchenpos.products.domain.tobe.domain;
+package kitchenpos.common.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import kitchenpos.FakeProfanities;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -12,7 +11,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 class DisplayedNameTest {
 
-    @DisplayName("상품의 이름에는 비속어가 포함될 수 없다.")
+    @DisplayName("이름에는 비속어가 포함될 수 없다.")
     @ValueSource(strings = {"비속어", "욕설"})
     @ParameterizedTest
     void 비속어(final String displayedName) {
@@ -21,19 +20,21 @@ class DisplayedNameTest {
         ).isInstanceOf(IllegalArgumentException.class);
     }
 
-    @DisplayName("상품 이름을 생성할 수 있다.")
-    @Test
-    void 생성() {
+    @DisplayName("이름을 생성할 수 있다.")
+    @ValueSource(strings = {"후라이드치킨", "양념치킨", "두마리메뉴"})
+    @ParameterizedTest
+    void 생성(final String displayedName) {
         assertDoesNotThrow(
-            () -> new DisplayedName("치킨", new FakeProfanities())
+            () -> new DisplayedName(displayedName, new FakeProfanities())
         );
     }
 
-    @DisplayName("상품 이름 간 동등성을 확인할 수 있다.")
+    @DisplayName("이름 간 동등성을 확인할 수 있다.")
     @Test
     void 동등성() {
         final DisplayedName displayedName1 = new DisplayedName("치킨", new FakeProfanities());
         final DisplayedName displayedName2 = new DisplayedName("치킨", new FakeProfanities());
+
         assertThat(displayedName1).isEqualTo(displayedName2);
     }
 }
